backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
cryptography==46.0.1
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
openpyxl==3.1.5
python-docx==1.2.0
pypdf==6.0.0
pytesseract==0.3.13
Pillow==11.3.0
email-validator==2.3.0
python-dateutil==2.9.0.post0
fpdf2==2.8.4
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.28.1

